requirements.txt
clickhouse-driver
python-binance